requirements.txt
# Core dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
termcolor>=2.3.0
asyncio>=3.4.3
pytz>=2023.3

# Database and data analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0

# NLP and text processing
nltk>=3.8.0
tqdm>=4.65.0

# Image and video processing
Pillow>=10.0.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
imageio>=2.30.0
imageio-ffmpeg>=0.4.8

# Web and API
requests>=2.30.0
aiohttp>=3.8.5

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
